Add files via upload
This project is a Machine Learning model that leverages historical health data to predict the likelihood of diabetes in individuals. By analyzing patterns and correlations in patient records, the algorithm can identify key predictors of diabetes and assess risk levels. This tool aims to support early diagnosis and preventive care, potentially reducing the long-term impact of diabetes on patients' lives. It's a testament to the power of AI in transforming healthcare outcomes.